Modify to avoid to overwirte git user setting

+++ install.bash
@@ -2,9 +2,11 @@
 
 dotfiles=${PWD}
 
-read -p "Please set your 'user name' on git : " git_user_name
-read -p "Please set your 'email' on git : " git_email
-echo -e "[user]\n\tname = ${git_user_name}\n\temail = ${git_email}\n" > ${dotfiles}/.gitconfig.user
+if [ ! -f ${dotfiles}/.gitconfig.user ]; then
+  read -p "Please set your 'user name' on git : " git_user_name
+  read -p "Please set your 'email' on git : " git_email
+  echo -e "[user]\n\tname = ${git_user_name}\n\temail = ${git_email}\n" > ${dotfiles}/.gitconfig.user
+fi
 
 ln -sfn  ${dotfiles}/.bashrc         ${HOME}/
 ln -sfn  ${dotfiles}/.gitconfig      ${HOME}/
